refactor(R): Renamed file with Cliffords r strategy.
