+++ src/simulation_strategies/cliffords_r_simulation_strategy.py
